+++ src/redux/actions/newsAction.ts
@@ -1,6 +1,5 @@
-import axios from "axios";
 import { Dispatch } from "redux";
-import { SET_ALL_STORIES, StoryType } from "../../utils/const";
+import { SET_ALL_STORIES} from "../../utils/const";
 
 //ЭКШЕН(или данные для экшена, хз). Запрашиваем данные всех постов по айди
 export const setStories = (data: any): any => {
@@ -10,29 +9,4 @@ export const setStories = (data: any): any => {
       payload: data,
     });
   };
-};
-//Promise<StoryType | StoryType[]> | undefined
-//запрашиваем ДАННЫЕ всех постов по айди
-export const getStoriesByIds = (ids:any): any => {
-  if (ids!==undefined) return Promise.all(ids.map((id: number) => getStoryById(id)));
-};
-//запрашиваем ДАННЫЕ всех НОВЫХ постов
-export const getStories = (): Promise<StoryType | StoryType[]> => {
-  return getNewStoriesIds().then((ids) =>
-    Promise.all(ids.map((id: number) => getStoryById(id)))
-  );
-};
-
-//запрашиваем ДАННЫЕ одного поста по одному айди
-export const getStoryById = (id: number): Promise<StoryType> => {
-  return axios
-    .get(`https://hacker-news.firebaseio.com/v0/item/${id}.json`)
-    .then((response) => response.data);
-};
-
-//запрашиваем АЙДИ 100 новых постов
-const getNewStoriesIds = (): Promise<Array<number>> => {
-  return axios
-    .get(`https://hacker-news.firebaseio.com/v0/newstories.json`)
-    .then((response) => response.data.slice(0, 100));
 };
