+++ src/styledComponents/Text.jsx
@@ -17,9 +17,26 @@ export const BoldTitle = styled.h2`
     font-size: 16px;
   }
 `;
+
+export const Text = styled.p`
+  font-family: Verdana, Geneva, sans-serif;
+  font-weight: 300;
+  margin:0px 5px 0px 5px;
+  @media (max-width: 480px) {
+    font-size: 14px;
+  }
+`;
+export const Author = styled(Text)`
+  font-weight: bold;
+`;
+export const Score = styled(Text)`
+  font-weight: bold;
+`;
+
 export const Description = styled.p`
   font-family: Verdana, Geneva, sans-serif;
   font-weight: 300;
+  margin:5px;
   //костыль всё, что ниже| на ссылки все равно не работает
   @media (max-width: 500px) { 
     font-size: 14px;
@@ -37,18 +54,4 @@ export const Description = styled.p`
     max-height:800px;
   }
   
-`;
-export const Text = styled.p`
-  font-family: Verdana, Geneva, sans-serif;
-  font-weight: 300;
-  margin:0px 5px 0px 5px;
-  @media (max-width: 480px) {
-    font-size: 14px;
-  }
-`;
-export const Author = styled(Text)`
-  font-weight: bold;
-`;
-export const Score = styled(Text)`
-  font-weight: bold;
 `;
